requirements.txt
anyio==4.6.2.post1
blinker==1.8.2
certifi==2024.8.30
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-telegram-bot==21.6
sniffio==1.3.1
Werkzeug==3.0.6
